chore(rocks_db): move ShallowTempDir to benches crate
